Move all survey-related classes to the survey package
