+++ app/src/androidTest/java/com/duckduckgo/app/survey/api/SurveyDownloaderTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018 DuckDuckGo
+ * Copyright (c) 2019 DuckDuckGo
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,16 +14,13 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.app.feedback.api
+package com.duckduckgo.app.survey.api
 
 import com.duckduckgo.app.InstantSchedulersRule
-import com.duckduckgo.app.feedback.db.SurveyDao
-import com.duckduckgo.app.feedback.model.Survey
-import com.duckduckgo.app.feedback.model.Survey.Status.NOT_ALLOCATED
-import com.duckduckgo.app.feedback.model.Survey.Status.SCHEDULED
-import com.duckduckgo.app.survey.api.SurveyDownloader
-import com.duckduckgo.app.survey.api.SurveyGroup
-import com.duckduckgo.app.survey.api.SurveyService
+import com.duckduckgo.app.survey.db.SurveyDao
+import com.duckduckgo.app.survey.model.Survey
+import com.duckduckgo.app.survey.model.Survey.Status.NOT_ALLOCATED
+import com.duckduckgo.app.survey.model.Survey.Status.SCHEDULED
 import com.nhaarman.mockitokotlin2.*
 import okhttp3.ResponseBody
 import org.junit.Rule
