+++ app/src/main/java/com/duckduckgo/app/feedback/ui/negative/subreason/SubReasonNegativeFeedbackFragment.kt
@@ -129,7 +129,7 @@ class SubReasonNegativeFeedbackFragment : FeedbackFragment() {
     }
 
     private fun appSlowOrBuggySubReasons(): List<FeedbackTypeSubReasonDisplay> {
-        return FeedbackType.PerformanceSubReasons.values().mapNotNull() {
+        return FeedbackType.PerformanceSubReasons.values().mapNotNull {
             FeedbackTypeDisplay.subReasons[it]
         }
     }
