+++ app/src/main/java/com/duckduckgo/app/survey/api/SurveyDownloader.kt
@@ -17,10 +17,10 @@
 package com.duckduckgo.app.survey.api
 
 import com.duckduckgo.app.survey.api.SurveyGroup.SurveyOption
-import com.duckduckgo.app.feedback.db.SurveyDao
-import com.duckduckgo.app.feedback.model.Survey
-import com.duckduckgo.app.feedback.model.Survey.Status.NOT_ALLOCATED
-import com.duckduckgo.app.feedback.model.Survey.Status.SCHEDULED
+import com.duckduckgo.app.survey.db.SurveyDao
+import com.duckduckgo.app.survey.model.Survey
+import com.duckduckgo.app.survey.model.Survey.Status.NOT_ALLOCATED
+import com.duckduckgo.app.survey.model.Survey.Status.SCHEDULED
 import io.reactivex.Completable
 import timber.log.Timber
 import java.io.IOException
