+++ app/src/main/res/layout/content_feedback.xml
@@ -50,7 +50,7 @@
             android:clickable="true"
             android:contentDescription="@string/feedbackInitialDisambiguationHappyButtonContentDescription"
             android:focusable="true"
-            tools:src="@drawable/ic_button_happy_light_theme"
+            tools:src="@drawable/button_happy_light_theme"
             app:layout_constraintEnd_toStartOf="@id/negativeFeedbackButton"
             app:layout_constraintHorizontal_chainStyle="packed"
             app:layout_constraintStart_toStartOf="parent"
@@ -62,7 +62,7 @@
             android:layout_height="100dp"
             android:background="?attr/selectableItemBackgroundBorderless"
             android:contentDescription="@string/feedbackInitialDisambiguationSadButtonContentDescription"
-            tools:src="@drawable/ic_button_sad_light_theme"
+            tools:src="@drawable/button_sad_light_theme"
             android:layout_marginStart="28dp"
             app:layout_constraintBottom_toBottomOf="@id/positiveFeedbackButton"
             app:layout_constraintEnd_toEndOf="parent"
